added tests for forging endpoint
