+++ test/api/delegates.js
@@ -6,7 +6,14 @@ var sendLISK = require('../common/complexTransactions.js').sendLISK;
 var sendTransaction = require('../common/complexTransactions.js').sendTransaction;
 var modulesLoader = require('./../common/initModule.js').modulesLoader;
 var genesisDelegates = require('../genesisDelegates.json');
+
 var testDelegate = genesisDelegates.delegates[0];
+
+function getForgingStatus (publicKey, cb) {
+	http.get('/api/delegates/forging/status?publicKey=' + publicKey, function (err, res) {
+		cb(err, res.body);
+	});
+}
 
 //insert one extra delegate
 before(function (done) {
@@ -788,19 +795,20 @@ describe('GET /api/delegates/forging/status', function () {
 	});
 });
 
-describe('POST /api/delegates/forging/disable', function () {
+describe('PUT /api/delegates/forging', function () {
 
 	before(function (done) {
 		http.get('/api/delegates/forging/status?publicKey=' + testDelegate.publicKey, function (err, res) {
 			node.expect(res.body).to.have.property('success').to.be.ok;
 			node.expect(res.body).to.have.property('enabled').to.be.a('boolean');
 			if (!res.body.enabled) {
-				http.post('/api/delegates/forging/enable', {
+				http.put('/api/delegates/forging', {
 					publicKey: testDelegate.publicKey,
 					key: testDelegate.key
 				}, function (err, res) {
 					node.expect(res.body).to.have.property('success').to.be.ok;
-					node.expect(res.body).to.have.property('address').equal(testDelegate.address);
+					node.expect(res.body).to.have.property('publicKey').equal(testDelegate.publicKey);
+					node.expect(res.body).to.have.property('forging').equal(true);
 					done();
 				});
 			} else {
@@ -810,7 +818,7 @@ describe('POST /api/delegates/forging/disable', function () {
 	});
 
 	it('using no params should fail', function (done) {
-		http.post('/api/delegates/forging/disable', {}, function (err, res) {
+		http.put('/api/delegates/forging', {}, function (err, res) {
 			node.expect(res.body).to.have.property('success').not.to.be.ok;
 			node.expect(res.body).to.have.property('error').to.be.a('string').and.to.contain('Missing required property: ');
 			done();
@@ -819,7 +827,8 @@ describe('POST /api/delegates/forging/disable', function () {
 
 	it('using invalid publicKey should fail', function (done) {
 		var invalidPublicKey= '9d3058175acab969f41ad9b86f7a2926c74258670fe56b37c429c01fca9fff0a';
-		http.post('/api/delegates/forging/disable', {
+
+		http.put('/api/delegates/forging', {
 			publicKey: invalidPublicKey,
 			key: testDelegate.key
 		}, function (err, res) {
@@ -830,7 +839,7 @@ describe('POST /api/delegates/forging/disable', function () {
 	});
 
 	it('using invalid key should fail', function (done) {
-		http.post('/api/delegates/forging/disable', {
+		http.put('/api/delegates/forging', {
 			publicKey: testDelegate.publicKey,
 			key: 'invalid key'
 		}, function (err, res) {
@@ -841,79 +850,28 @@ describe('POST /api/delegates/forging/disable', function () {
 	});
 
 	it('using valid params should be ok', function (done) {
-		http.post('/api/delegates/forging/disable', {
+		http.put('/api/delegates/forging', {
+			key: testDelegate.key,
 			publicKey: testDelegate.publicKey,
-			key: testDelegate.key
 		}, function (err, res) {
-			node.expect(res.body).to.have.property('success').to.be.ok;
-			node.expect(res.body).to.have.property('address').equal(testDelegate.address);
+			node.expect(res.body).to.have.property('publicKey').equal(testDelegate.publicKey);
+			node.expect(res.body).to.have.property('forging').to.be.a('boolean');
 			done();
 		});
 	});
-});
 
-describe('POST /api/delegates/forging/enable', function () {
-	var testDelegate = genesisDelegates.delegates[0];
+	it('using valid params should toggle forging status', function (done) {
+		getForgingStatus(testDelegate.publicKey, function (err, res) {
+			var currentStatus = res.enabled;
 
-	before(function (done) {
-		http.get('/api/delegates/forging/status?publicKey=' + testDelegate.publicKey, function (err, res) {
-			node.expect(res.body).to.have.property('success').to.be.ok;
-			node.expect(res.body).to.have.property('enabled').to.be.a('boolean');
-			if (res.body.enabled) {
-				http.post('/api/delegates/forging/disable', {
-					publicKey: testDelegate.publicKey,
-					key: testDelegate.key
-				}, function (err, res) {
-					node.expect(res.body).to.have.property('success').to.be.ok;
-					node.expect(res.body).to.have.property('address').equal(testDelegate.address);
-					done();
-				});
-			} else {
+			http.put('/api/delegates/forging', {
+				publicKey: testDelegate.publicKey,
+				key: testDelegate.key
+			}, function (err, res) {
+				node.expect(res.body).to.have.property('publicKey').equal(testDelegate.publicKey);
+				node.expect(res.body).to.have.property('forging').to.not.equal(currentStatus);
 				done();
-			}
-		});
-	});
-
-	it('using no params should fail', function (done) {
-		http.post('/api/delegates/forging/enable', {}, function (err, res) {
-			node.expect(res.body).to.have.property('success').not.to.be.ok;
-			node.expect(res.body).to.have.property('error').to.be.a('string').and.to.contain('Missing required property: publicKey');
-			done();
-		});
-	});
-
-	it('using invalid key should fail', function (done) {
-		http.post('/api/delegates/forging/enable', {
-			publicKey: testDelegate.publicKey,
-			key: 'invalid key'
-		}, function (err, res) {
-			node.expect(res.body).to.have.property('success').not.to.be.ok;
-			node.expect(res.body).to.have.property('error').to.be.a('string').and.to.contain('Invalid key and public key combination');
-			done();
-		});
-	});
-
-	it('using invalid publicKey should fail', function (done) {
-		var invalidPublicKey= '9d3058175acab969f41ad9b86f7a2926c74258670fe56b37c429c01fca9fff0a';
-		http.post('/api/delegates/forging/enable', {
-			publicKey: invalidPublicKey,
-			key: testDelegate.key
-		}, function (err, res) {
-			node.expect(res.body).to.have.property('success').not.to.be.ok;
-			node.expect(res.body).to.have.property('error').to.be.a('string').and.to.contain(['Delegate with publicKey:', invalidPublicKey, 'not found'].join(' '));
-			done();
-		});
-	});
-
-
-	it('using valid params should be ok', function (done) {
-		http.post('/api/delegates/forging/enable', {
-			publicKey: testDelegate.publicKey,
-			key: testDelegate.key
-		}, function (err, res) {
-			node.expect(res.body).to.have.property('success').to.be.ok;
-			node.expect(res.body).to.have.property('address').equal(testDelegate.address);
-			done();
+			});
 		});
 	});
 });
